Fix compilation error: cast int to short for head field

+++ src/main/java/Exe_Z/bot/player/BotPlayer.java
@@ -52,7 +52,7 @@ public class BotPlayer {
             botChar.mp = botChar.maxMP;
             botChar.isDead = false;
             if (botChar.head < 0) {
-                botChar.head = botChar.gender == 0 ? 28 : 26;
+                botChar.head = (short) (botChar.gender == 0 ? 28 : 26);
             }
             int map = botChar.mapId;
             int zoneId = NinjaUtils.randomZoneId(map);
